#57 Added UnitTest for WhenOrNone
